Relecture FR Boris
Some proposals, take what you want :)

A note: French people who are concerned with gender equality tend to use specific syntaxes in their writing. For example, instead of "développeurs", we will write "développeurs et développeuses" or, in a synthetic form, "développeurs·euses" (or [a lot of things](https://codepen.io/vincent-valentin/full/woGLVL), as the usage of gender-neutral French is quite new). I don't know if you want to do (and nobody will expect it from someone who's not fluent) but I thought you would appreciate to know.

Another note: As you might expect, there is a real separation among French developers between those who are interested in a software craftmanship that is very disconnected from the user experience and those who are looking to produce a high-quality web for their visitors. The latter designate this knowledge under the label "Qualité Web" and one of the biggest players in France is Opquast, an organization that trains in Web quality, and offers checklists for all types of audiences. Here is an EN translation of their main checklist: 
https://checklists.opquast.com/en/qualiteweb/. Elie Sloim, the face of Opquast, was a speaker at [CSUN 2018](https://www.slideshare.net/ElieSlom/web-quality-management-improving-ux-for-everyone-91609245).